fixes for GCC, stack trace is broken (will fix in the next commit)

+++ source/kibble/platform/arch.cpp
@@ -83,7 +83,7 @@ CPUInfo::Internal::Internal()
     }
 
     // Get highest extended function ID
-    cpuid(cpui.data(), 0x80000000L);
+    cpuid(cpui.data(), int32_t(0x80000000L));
     unsigned int n_ex_ids = static_cast<unsigned int>(cpui[0]);
 
     char brand[0x40];
